requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
python-multipart
pytest
httpx
pydantic
dotenv
email-validator>=2.0.0